Add autoJoinApprover to control auto-joining behavior
At first, I was considering just adding `autojoin: true|false`, but
that's not very flexible and external libraries will have to:

- either use it and auto-join all rooms (`autojoin: true`) - not very
  useful if you want more control
- or, don't use it at all (`autojoin: false`) and handle joins or
  invitation-rejection completely by yourself - not very helpful

We now have an `autoJoinApprover` callback, so the consuming library
only needs to provide that and have an easy say on whether joining
should happen or not. The actual joining (or rejecting the invitation)
is still handled by linkpearl, along with all the retry logic, etc.

`autoleave` may be redone in a similar way for consistency and for
greater control.